separate folders and remove files
