+++ .gitignore
@@ -1,0 +1,10 @@
+*.o
+*.a
+*.exe
+*.out
+test
+*.rpt
+*.dSYM
+.DS_Store
+.vscode/*
+.gitattributes
