+++ backend/run.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+# Move to the directory
+cd /code/backend/
+pip install --upgrade pip
+
+cd /code/backend/
+
+# run server
+python manage.py runserver 0.0.0.0:8000
