+++ nginx/nginx.conf
@@ -1,0 +1,1 @@
+#for keep the folder
