--- user_management/accounts/serializers.py
@@ -1,10 +1,0 @@
-from rest_framework import serializers
-from .models import User
-
-
-class UserSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = User
-        fields = ('id', 'username', 'first_name', 'last_name', 'bio', 'profile_pic')
-        read_only_fields = ('username', )
