--- user_management/docker-compose.yml
@@ -1,42 +1,0 @@
-# version: '3'
-
-services:
-  db:
-    image: postgres
-    volumes:
-      - postgres-data:/var/lib/postgresql/data
-    environment:
-      - POSTGRES_DB=postgres
-      - POSTGRES_USER=postgres
-      - POSTGRES_PASSWORD=postgres
-
-  web:
-    build: .
-    command: python manage.py runserver 0.0.0.0:8000
-    volumes:
-      - .:/code
-    ports:
-      - "8000:8000"
-    environment:
-      - DB_NAME=postgres
-      - DB_USER=postgres
-      - DB_PASSWORD=postgres
-      - DB_HOST=db
-      - DB_PORT=5432
-    depends_on:
-      - db
-  
-  # nginx:
-  #   image: nginx:latest
-  #   ports:
-  #     - "80:80"
-  #   volumes:
-  #     - ./nginx.conf:/etc/nginx/nginx.conf:ro
-  #     - ./modsecurity.conf:/etc/nginx/modsecurity.conf:ro
-  #     - ./owasp-modsecurity-crs:/etc/nginx/owasp-modsecurity-crs:ro
-  #   depends_on:
-  #     - web
-
-volumes:
-  postgres-data:
-
